editor모드 일때 form 구현 완료
