+++ src/pages/MyWeki/components/MyWekiUI.tsx
@@ -1,8 +1,9 @@
-import React from 'react';
+import React, { useState } from 'react';
+import Cookies from 'js-cookie';
 import { profileType } from '@api/profile';
 import Button from '@components/button/Button';
+import EditorUI from './EditorUI';
 import ProfileUI from './ProfileUI';
-import WritingUI from './WritingUI';
 import { StyledMarkUpWrap } from './markUp.styled';
 import {
   StyledWekiWrap,
@@ -12,27 +13,40 @@ import {
 
 interface MyeWekiUIProps {
   profile: profileType | null;
-  isEdit?: boolean;
+  isEdit: boolean;
+  setIsEdit: React.Dispatch<React.SetStateAction<boolean>>;
 }
-const MyWekiUI = ({ profile, isEdit = true }: MyeWekiUIProps) => {
+const MyWekiUI = ({ profile, isEdit, setIsEdit }: MyeWekiUIProps) => {
   return (
     <StyledWekiWrap>
       <StyledWekiContent>
-        <StyledWekiHeader>
-          <div>
-            <h2>{profile?.name}</h2>
-            <Button $primary $width="160px" $height="45px">
-              위키참여하기
-            </Button>
-          </div>
-        </StyledWekiHeader>
+        {!isEdit && (
+          <StyledWekiHeader>
+            <div>
+              <h2>{profile ? profile.name : ''}</h2>
+              <Button $primary $width="160px" $height="45px">
+                위키참여하기
+              </Button>
+            </div>
+          </StyledWekiHeader>
+        )}
+
         <StyledMarkUpWrap>
           {!isEdit && (
             <div
-              dangerouslySetInnerHTML={{ __html: profile?.content as string }}
+              dangerouslySetInnerHTML={{
+                __html: profile ? profile.content : '',
+              }}
             ></div>
           )}
-          {isEdit && <WritingUI content={profile?.content as string} />}
+
+          {isEdit && Cookies.get('aceessToken') !== 'dad' && (
+            <EditorUI
+              setIsEdit={setIsEdit}
+              content={profile?.content as string}
+              name={profile ? profile.name : ''}
+            />
+          )}
         </StyledMarkUpWrap>
       </StyledWekiContent>
       {!isEdit && <ProfileUI {...profile} />}
