+++ src/pages/MyWeki/components/markUp.styled.ts
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import { theme } from '@styles/theme';
 
 export const StyledMarkUpWrap = styled.div`
   margin-top: 50px;
